remove some experiment-specific runnable functionality
